Add Maven framework
