+++ DM Course work/src/main/java/ru/shemplo/dm/course/RunCourseWork.java
@@ -1,0 +1,18 @@
+package ru.shemplo.dm.course;
+
+import javafx.application.Application;
+import javafx.stage.Stage;
+
+public class RunCourseWork extends Application {
+    
+    public static void main (String ... args) {
+        
+    }
+
+    @Override
+    public void start (Stage stage) throws Exception {
+        stage.setTitle ("Course work");
+        stage.show ();
+    }
+    
+}
