+++ DM Course work/src/main/java/ru/shemplo/dm/course/gfx/ProgressListener.java
@@ -1,0 +1,12 @@
+package ru.shemplo.dm.course.gfx;
+
+
+public interface ProgressListener {
+    
+    public void onComputationStarted ();
+    
+    public void onStepComputed ();
+    
+    public void onComputationFinished ();
+    
+}
