+++ DM Course work/src/main/java/ru/shemplo/dm/course/gfx/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * 
+ */
+/**
+ * @author Shemplo
+ *
+ */
+package ru.shemplo.dm.course.gfx;
