--- DM Course work/src/ru/shemplo/dm/course/gfx/ProgressListener.java
@@ -1,12 +1,0 @@
-package ru.shemplo.dm.course.gfx;
-
-
-public interface ProgressListener {
-    
-    public void onComputationStarted ();
-    
-    public void onStepComputed ();
-    
-    public void onComputationFinished ();
-    
-}
