--- DM Course work/src/ru/shemplo/dm/course/gfx/package-info.java
@@ -1,8 +1,0 @@
-/**
- * 
- */
-/**
- * @author Shemplo
- *
- */
-package ru.shemplo.dm.course.gfx;
